requirements.txt
torch
numpy
matplotlib
pillow
gradio
saliency
transformers